requirements.txt
pr_agent==0.2.4
transformers==4.44.0
fastembed==0.3.4
huggingface_hub==0.24.5
unstructured[local-inference]==0.15.5
sentence-transformers==3.0.1
python-dotenv==1.0.1
einops==0.8.0
llama-index==0.11.3
llama-index-llms-openai-like==0.2.0
llama-index-readers-github==0.2.0
llama-index-readers-web==0.2.2
llama-index-vector-stores-qdrant==0.3.0
llama-index-embeddings-huggingface==0.3.1
setuptools==72.2.0
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
PyGithub==1.59.1
tqdm==4.66.5
aiohttp==3.10.5